requirements.txt
numpy
pandas
scikit-learn
torch
pytorch-lightning
pyyaml
matplotlib
reportlab
